refactor: reorganize and update DHCP schemas and imports across modules
