+++ app/ldap_protocol/dhcp/__init__.py
@@ -12,17 +12,6 @@ from .exceptions import (
 )
 from .kea_dhcp_manager import KeaDHCPManager
 from .kea_dhcp_repository import KeaDHCPAPIRepository
-from .schemas import (
-    DHCPChangeStateSchemaRequest,
-    DHCPLeaseSchemaRequest,
-    DHCPLeaseSchemaResponse,
-    DHCPLeaseToReservationErrorResponse,
-    DHCPReservationSchemaRequest,
-    DHCPReservationSchemaResponse,
-    DHCPStateSchemaResponse,
-    DHCPSubnetSchemaAddRequest,
-    DHCPSubnetSchemaResponse,
-)
 from .stub import StubDHCPAPIRepository, StubDHCPManager
 
 
@@ -58,13 +47,4 @@ __all__ = [
     "DHCPOperationError",
     "DHCPAPIError",
     "DHCPSubnetSchemaRequest",
-    "DHCPSubnetSchemaAddRequest",
-    "DHCPReservationSchemaRequest",
-    "DHCPSubnetSchemaResponse",
-    "DHCPLeaseSchemaRequest",
-    "DHCPLeaseSchemaResponse",
-    "DHCPReservationSchemaResponse",
-    "DHCPChangeStateSchemaRequest",
-    "DHCPStateSchemaResponse",
-    "DHCPLeaseToReservationErrorResponse",
 ]
